+++ src/Footer/Footer.jsx
@@ -1,0 +1,9 @@
+const Footer = () => {
+    return (
+        <footer style={{padding: '20px', color: 'white', backgroundColor: 'blue'}}>
+            فوتر سایت
+        </footer>
+    )
+}
+
+export default Footer;
