+++ src/Products/Products.jsx
@@ -1,118 +1,48 @@
 import { useEffect, useState } from 'react';
-
-const CATEGORY_LAPTOP_ID =  'laptop'
-const CATEGORY_LAPTOP_LABEL = "لپ تاپ"
-const CATEGORY_MOBILE_ID = "mobile"
-const CATEGORY_MOBILE_LABEL = "موبایل"
-const CATEGORY_CAMERA_ID = "camera"
-const CATEGORY_CAMERA_LABEL = "دوربین"
-
-const CATEGORY_FOOD_ID = "food"
-const CATEGORY_FOOD_LABEL = "غذا"
-
-const CATEGORIES = [
-    {id: CATEGORY_LAPTOP_ID, label: CATEGORY_LAPTOP_LABEL},
-    {id: CATEGORY_MOBILE_ID, label: CATEGORY_MOBILE_LABEL},
-    {id: CATEGORY_CAMERA_ID, label: CATEGORY_CAMERA_LABEL},
-    {id: CATEGORY_FOOD_ID, label: CATEGORY_FOOD_LABEL},
-]
-
-const arrayOfProducts = [
-    {
-        id: 1,
-        name: "مک بوک پرو",
-        price: 1500,
-        category: CATEGORY_LAPTOP_ID
-    },
-    {
-        id: 2,
-        name: "مک بوک پرو دو",
-        price: 800,
-        category: CATEGORY_LAPTOP_ID
-    },
-    {
-        id: 3,
-        name: "مک بوک ایر",
-        price: 2000,
-        category: CATEGORY_LAPTOP_ID
-    },
-    {
-        id: 4,
-        name: "موبایل ایفون شش",
-        price: 900,
-        category: CATEGORY_MOBILE_ID
-    },
-    {
-        id: 5,
-        name: "موبایل ایفون پونزده پرو",
-        price: 2500,
-        category: CATEGORY_MOBILE_ID
-    },
-    {
-        id: 6,
-        name: "موبایل شیائومی",
-        price: 1000,
-        category: CATEGORY_MOBILE_ID
-    },
-    {
-        id: 7,
-        name: "لپ تاپ ایسوس",
-        price: 3000,
-        category: CATEGORY_LAPTOP_ID
-    },
-    {
-        id: 8,
-        name: "لپ تاپ خیلی گرون",
-        price: 3000,
-        category: CATEGORY_LAPTOP_ID
-    },
-    {
-        id: 9,
-        name: "موبایل سامسونگ",
-        price: 3000,
-        category: CATEGORY_MOBILE_ID
-    },
-    {
-        id: 10,
-        name: "دوربین کانون",
-        price: 1200,
-        category: CATEGORY_CAMERA_ID
-    },
-    {
-        id: 11,
-        name: "دوربین فوجی",
-        price: 1300,
-        category: CATEGORY_CAMERA_ID
-    }
-]
-
-
-const Input = (props) => (
-        <input
-            style={{border: '1px solid red', padding: '10px', borderRadius: '5px'}}
-            {...props}
-        />
-    )
+import Input from '../components/Input';
 
 const Products = () => {
     const [name, setName] = useState('');
-    const [price, setPrice] = useState(100);
+    const [price, setPrice] = useState(0);
     const [category, setcategory] = useState('');
 
-    const findCategoryById = (id) => {
-        return CATEGORIES.find(item => item.id === id)
-    }
+    const [categories, setCategories] = useState([]);
+    const [loading, setLoading] = useState(true);
+    const [products, setProducts] = useState();
 
-    const filteredProducts = arrayOfProducts.filter((item) => 
-        item.name.includes(name) && 
-        item.price > price &&
-        item.category.includes(category)
-    )
+    console.log({
+        products, loading
+    });
 
     useEffect(() => {
-        setPrice('');
-        setName('');
-    }, [category])
+        fetch('https://dummyjson.com/products')
+            .then(res => res.json())
+            .then(res => {
+                setProducts(res.products)
+
+                // TODO:
+                setCategories([])
+                setLoading(false)
+            })
+    }, []);
+
+    const findCategoryById = (id) => {
+        // TODO:
+        return [].find(item => item.id === id)
+    }
+
+    if (loading) {
+        return <div>Loading...</div>
+    }
+
+    if (!products) {
+        return <div>No data</div>
+    }
+
+    const filteredProducts = products.filter(item => {
+        return item.title.toLowerCase().includes(name.toLowerCase()) &&
+                item.price >= price
+    })
 
     return (
         <div>
@@ -132,14 +62,11 @@ const Products = () => {
             <select onChange={e => {
                 setcategory(e.target.value)
                 }}>
-                {CATEGORIES.map(category => (
+                {categories.map(category => (
                     <option key={category.id} value={category.id}>{category.label}</option>
                 ))}
-            </select>
-            <button onClick={() => setcategory(CATEGORY_FOOD_ID)}>ریست کردن غذا</button>
-            <ul>
-                {filteredProducts.map(item => <li key={item.id}>{item.name}: {item.price} {findCategoryById(item.category).label}</li>)}
-            </ul>
+            </select>            
+            {filteredProducts.map(item => <li key={item.id}>{item.title} {item.price}</li>)}
         </div>
     )
 }
