+++ src/components/Input.jsx
@@ -1,0 +1,8 @@
+const Input = (props) => (
+    <input
+        style={{border: '1px solid red', padding: '10px', borderRadius: '5px'}}
+        {...props}
+    />
+)
+
+export default Input;
